% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{get_optimal_pcs}
\alias{get_optimal_pcs}
\title{Get Optimal Number of PCs}
\usage{
get_optimal_pcs(seurat_obj, reduction = "pca")
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{reduction}{Type of reduction (e.g., 'pca')}
}
\value{
Number of PCs to use
}
\description{
Get the optimal number of principal components to use
}
