% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{multi_f_plots}
\alias{multi_f_plots}
\title{Plot multiple features with color map}
\usage{
multi_f_plots(
  seurat_obj,
  feature_list,
  ncol = 3,
  pt.size = 0.8,
  split_by = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{feature_list}{List of features to plot}

\item{ncol}{Number of columns for arrangement}

\item{pt.size}{Point size for plotting}

\item{split_by}{Variable for splitting}
}
\value{
Plot grid
}
\description{
Plot multiple features with custom color scale
}
