% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_compare_normalisation}
\alias{plot_compare_normalisation}
\title{plot_compare_normalisation}
\usage{
plot_compare_normalisation(seuratObj)
}
\arguments{
\item{seuratObj}{A Seurat object with both RNA and SCT assays}
}
\value{
A plot comparing raw counts, normalized counts, SCTransform counts, and scaled data
}
\description{
Function to compare different normalization methods
}
