% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_multi_feat_umap}
\alias{plot_multi_feat_umap}
\title{Plot UMAPs for various feature sets}
\usage{
plot_multi_feat_umap(
  seurat_obj,
  features,
  min.dist = 0.1,
  max_features = 3000,
  ftype = "HVG",
  step = 500,
  out_name = "combined_umap"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{features}{List of features to use}

\item{min.dist}{Minimum distance parameter for UMAP}

\item{max_features}{Maximum number of features to consider}

\item{ftype}{Type of features (e.g., 'HVG')}

\item{step}{Step size for incrementing feature sets}

\item{out_name}{Name to assign to the combined plot}
}
\value{
Combined UMAP plot
}
\description{
Plot UMAPs for different numbers of features (e.g., Highly Variable Genes or Most Abundant Genes)
}
