% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_multi_min_dist_umap}
\alias{plot_multi_min_dist_umap}
\title{Plot UMAPs to test min.dist}
\usage{
plot_multi_min_dist_umap(
  object = seu_obj,
  features = NULL,
  dims = 1:30,
  out_name = "min_dist_umaps"
)
}
\value{
plot saved to global environment
}
\description{
plots multiple UMAP's at different min.dist values ...
}
