% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{qc}
\alias{qc}
\title{Seurat Object Quality Control function}
\usage{
qc(
  matrix,
  nf_min = nf_min,
  nf_max = nf_max,
  nc = nc,
  pmt = pmt,
  group = "orig.ident",
  colors = NULL
)
}
\arguments{
\item{matrix}{Seurat object ...}

\item{nf_min}{Minimum number of features ...}

\item{nf_max}{Maximum number of features ...}

\item{nc}{Maximum number of counts ...}

\item{pmt}{Percentage of mitochondrial genes ...}

\item{group}{Grouping variable ...}

\item{colors}{Colors for facetting ...}
}
\value{
Seurat object after quality control
}
\description{
Run Iteratively for the necessary QC...
}
