% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{scds_doublets}
\alias{scds_doublets}
\title{scds Doublet finder for Seurat}
\usage{
scds_doublets(matrix)
}
\arguments{
\item{matrix}{Seurat object ...}
}
\value{
Seurat object with doublet scores and call
}
\description{
Use not recommended ...
}
