% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie3DFeaturePlot}
\alias{wookie3DFeaturePlot}
\title{Plot 3D Umap}
\usage{
wookie3DFeaturePlot(
  seu,
  gene = NULL,
  palette = viridis::viridis(50),
  marker.size = 2,
  reduction = "umap",
  silentwookie = TRUE,
  opacity = 1
)
}
\arguments{
\item{seu}{Seurat Object}

\item{palette}{color palette to plot, Default is NULL}

\item{marker.size}{default 2}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}

\item{opacity}{default 1}

\item{group.by}{metadata column to group by}
}
\value{
plot
}
\description{
Plot 3D UMAP
}
