% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_annotate}
\alias{wookie_annotate}
\title{Function to plot cluster Similarity}
\usage{
wookie_annotate(object, marker_gene_list, threshold = 0)
}
\arguments{
\item{object}{Seurat object}

\item{marker_gene_list}{a nested list of marker genes,sub list names are used as labels}

\item{threshold}{confidence score threshold to label a cell}
}
\value{
object
}
\description{
Function to annotate cells given a nested list of marker genes,sub list names are used as labels
}
