% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_batch_entropy}
\alias{wookie_batch_entropy}
\title{Function to plot entropy of a batch}
\usage{
wookie_batch_entropy(
  seurat_obj,
  batch.key = "batch",
  cluster.key = "seurat_clusters"
)
}
\arguments{
\item{batch.key}{metadata column with batch information}

\item{cluster.key}{metadata column with cluster information}

\item{object}{Seurat object}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
plot
}
\description{
Function to plot entropy of a batch across clusters to evaluate mixing between batches, higher the entropy better the mixing
}
