% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_clusterSimilarityPlot}
\alias{wookie_clusterSimilarityPlot}
\title{Function to plot cluster Similarity}
\usage{
wookie_clusterSimilarityPlot(
  seurat_obj,
  dims = 1:30,
  clusters = "seurat_clusters",
  reduction = "PCA",
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{dims}{dimension, default is 1:30}

\item{clusters}{default is 'seurat_clusters'}

\item{reduction}{default is 'PCA'}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
cluster Similarity heatmap
}
\description{
Function to plot cluster Similarity
}
