% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_evaluate_batch_correction}
\alias{wookie_evaluate_batch_correction}
\title{Function to plot shared neighbours per batch}
\usage{
wookie_evaluate_batch_correction(
  seurat_obj,
  reduction = "harmony",
  batch_col = NULL,
  group_by_cluster = FALSE,
  cluster_col = NULL,
  k.param = 20,
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{reduction}{default is 'harmony'}

\item{batch_col}{metadata column with batch information}

\item{group_by_cluster}{set to TRUE, if the plot should be grouped by clusters. default is FALSE}

\item{cluster_col}{metadata column with cluster information to group by}

\item{k.param}{k.param for Seurat::FindNeighbours}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
plot
}
\description{
Function to plot shared neighbours per batch, to evaluate batch correction, 
i.e  the idea being that each cell should have a similar number of shared nearest neighbours from each batch.
}
