% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_fc_hist}
\alias{wookie_fc_hist}
\title{histograms of nFeature and nCounts}
\usage{
wookie_fc_hist(seurat_obj, title = "Histogram", fi = 0, ci = 0)
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{title}{title of the plot}

\item{fi}{features threshold}

\item{ci}{counts threshold}
}
\value{
plot
}
\description{
Get histograms of nFeature and nCounts and possible thresholds
}
