% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_featureplot}
\alias{wookie_featureplot}
\title{Plot multiple features with color map}
\usage{
wookie_featureplot(
  seuratObject,
  featureList,
  ncol = 3,
  pt.size = 0.8,
  split_by = NULL
)
}
\arguments{
\item{ncol}{Number of columns for arrangement}

\item{pt.size}{Point size for plotting}

\item{split_by}{Variable for splitting}

\item{seurat_obj}{Seurat object}

\item{feature_list}{List of features to plot}
}
\value{
Plot grid
}
\description{
Plot multiple features with custom color scale
}
