% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_filter_celltype}
\alias{wookie_filter_celltype}
\title{Filter particular cell types based on marker gene expression}
\usage{
wookie_filter_celltype(seurat_obj, marker_list, cutoff = 0.99)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{marker_list}{List of marker genes for the cell type to remove}

\item{cutoff}{Quantile threshold (default: 0.99)}
}
\value{
Filtered Seurat object
}
\description{
Remove specific cell types based on marker gene expression
}
