% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_filter_celltypes}
\alias{wookie_filter_celltypes}
\alias{wookie_filter_celltype}
\title{wookie_filter_celltypes}
\usage{
wookie_filter_celltype(seurat_object, marker_list, cutoff = 0.99)
}
\arguments{
\item{seurat_object}{object to filter}

\item{cutoff}{quantile threshold, default is 0.99}

\item{marker_List}{list of marker genes of celltype to remove ...}
}
\value{
filtered seurat object
}
\description{
Function to remove particular cell types based on marker gene expression ...
}
