% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_find_resolution}
\alias{wookie_find_resolution}
\title{Function to plot number of clusters per resolution}
\usage{
wookie_find_resolution(
  seurat_obj,
  algorithm = 1,
  resolution_range = c(0.1, 2),
  resolution_step = 0.01
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{algorithm}{same as Seurat FindClusters, Default is 1 : Louvain}

\item{resolution_range}{range of resolution to plot, default is : c(0.1, 2)}

\item{resolution_step}{default is 0.01}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
plot
}
\description{
Function to plot number of clusters across a set of resolutions, to identify an optimal resolution
}
