% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_ge_histogram}
\alias{wookie_ge_histogram}
\title{Compare Log Normalisation and SCT (Histogram)}
\usage{
wookie_ge_histogram(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat Object with both RNA and SCT assays}
}
\value{
Histogram comparing Log Normalisation and SCT
}
\description{
Compare Log Normalisation and SCT
}
