% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_get_filters}
\alias{wookie_get_filters}
\title{Function to get qc filters}
\usage{
wookie_get_filters(seurat_obj, multiplier = 2)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{multiplier}{default is 2 i.e 2 times SD}
}
\description{
Function to get filters, to get data which lies between the mean and +/- multiplier value times SD of the distribution
}
