% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_matrix_qc_plot}
\alias{wookie_matrix_qc_plot}
\title{Plot QC metrics of a sparse matrix}
\usage{
wookie_matrix_qc_plot(count_matrix_sparse, fill_color = "#589FFF", title = "")
}
\arguments{
\item{count_matrix_sparse}{Sparse matrix}

\item{fill_color}{Plot color (default: #589FFF)}

\item{title}{Title for the plot}
}
\value{
QC plot
}
\description{
Plot QC metrics of a sparse matrix
}
