% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_matrix_sum}
\alias{wookie_matrix_sum}
\title{Matrix Sum function}
\usage{
wookie_matrix_sum(
  matrix1,
  matrix2,
  sample = "sample",
  min_cells = 3,
  min_features = 200,
  silentwookie = FALSE
)
}
\arguments{
\item{matrix1}{count matrix ...}

\item{matrix2}{count matrix ...}

\item{sample}{sample/library name ...}

\item{min_cells}{minimum number cells a gene is found in ...}

\item{min_features}{minimum number of features found in a cell ...}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
Summed and merged Seurat object
}
\description{
Merge two count matrices, where the cells are the same, to obtain a single seurat object with the counts from two matrices summed for each cell ...
}
