% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_Mindist}
\alias{wookie_Mindist}
\title{Plot UMAPs for various min.dist values}
\usage{
wookie_Mindist(
  seurat_obj,
  features = NULL,
  dims = 1:30,
  out_name = "min_dist_umaps",
  reduction = "pca",
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{features}{Features to use for UMAP}

\item{dims}{Dimensions to use for UMAP}

\item{out_name}{Name to assign to the combined plot}

\item{reduction}{reduction for RunUMAP , deafult is 'pca'}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
Combined UMAP plot
}
\description{
Plot UMAPs for different min.dist values
}
