% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_scrub}
\alias{wookie_scrub}
\title{Scrublet for Seurat ...}
\usage{
wookie_scrub(seu_obj, preprocess = FALSE)
}
\arguments{
\item{seu_obj}{Seurat object ...}

\item{preprocess}{specify if the object has been preprocces and runPCA was done  ...}
}
\value{
Seurat object with scrublet scores and call
}
\description{
Run Scrublet on a Seurat Object ...
}
