% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_silhouettePlot}
\alias{wookie_silhouettePlot}
\title{Function to plot silhouette scores for each cluster}
\usage{
wookie_silhouettePlot(
  seurat,
  cluster = "seurat_clusters",
  dims = 1:30,
  reduction = "pca"
)
}
\arguments{
\item{seurat}{Seurat object}

\item{cluster}{default is 'seurat_clusters'}

\item{dims}{dimension, default is 1:30}

\item{reduction}{default is 'pca'}
}
\value{
silhouette plot
}
\description{
Function to plot silhouette scores for each cluster
}
