% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_silhouttePlot}
\alias{wookie_silhouttePlot}
\title{Function to plot silhoutte scores for each cluster}
\usage{
wookie_silhouttePlot(
  seurat,
  cluster = "seurat_clusters",
  dims = 1:30,
  reduction = "pca"
)
}
\arguments{
\item{seurat}{Seurat object}

\item{cluster}{default is 'seurat_clusters'}

\item{dims}{dimension, default is 1:30}

\item{reduction}{default is 'pca'}
}
\value{
silhoutte plot
}
\description{
Function to plot silhoutte scores for each cluster
}
