% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_stats}
\alias{wookie_stats}
\title{Function to get qc stats}
\usage{
wookie_stats(
  seurat_obj,
  counts = "nCount_RNA",
  features = "nFeature_RNA",
  mito = "percent.mt",
  ribo = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{counts}{default is 'nCount_RNA'}

\item{features}{default is 'nFeature_RNA'}

\item{mito}{default is 'percent.mt'}

\item{ribo}{default is NULL}
}
\value{
stats df
}
\description{
Function to get qc stats
}
