% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieBatchEntropy}
\alias{wookieBatchEntropy}
\title{Function to plot entropy of a batch}
\usage{
wookieBatchEntropy(
  seurat_obj,
  batch.key = "batch",
  cluster.key = "seurat_clusters",
  silentwookie = FALSE
)
}
\arguments{
\item{batch.key}{metadata column with batch information}

\item{cluster.key}{metadata column with cluster information}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}

\item{object}{Seurat object}
}
\value{
plot
}
\description{
Function to plot entropy of a batch across clusters to evaluate mixing between batches, higher the entropy better the mixing
}
