% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieClustertree}
\alias{wookieClustertree}
\title{Function to plot Cluster relationship Tree}
\usage{
wookieClustertree(
  seurat_obj,
  dims = NULL,
  features = NULL,
  reorder = FALSE,
  reorder.numeric = FALSE,
  saveplot = FALSE,
  assay = "RNA",
  reduction = "pca",
  slot = "data",
  dpi = 1080,
  height = 10,
  width = 10,
  units = "cm",
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{dims}{dimensions to use}

\item{reorder}{default is FALSE}

\item{reorder.numeric}{default is FALSE}

\item{saveplot}{save a jpeg image to working directory}

\item{assay}{default is 'RNA'}

\item{reduction}{default is 'pca'}

\item{slot}{default is 'data'}

\item{dpi}{default is 1080}

\item{height}{default is 10}

\item{width}{default is 10}

\item{units}{deafults is 'cm'}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
p
}
\description{
Function to plot Cluster relationship Tree
}
