% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieDotplot}
\alias{wookieDotplot}
\title{Plot dotplots with color map}
\usage{
wookieDotplot(
  seurat_obj,
  feature_list,
  assay = "RNA",
  scale.by = "size",
  tag = "DotPlot",
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{feature_list}{List of markers}

\item{assay}{default RNA}

\item{scale.by}{scale.by}

\item{tag}{plot title}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
Pdotplot
}
\description{
dotplot
}
