% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieFilterCelltype}
\alias{wookieFilterCelltype}
\title{Filter particular cell types based on marker gene expression}
\usage{
wookieFilterCelltype(
  seurat_obj,
  marker_list,
  cutoff = 0.99,
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{marker_list}{List of marker genes for the cell type to remove}

\item{cutoff}{Quantile threshold (default: 0.99)}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
Filtered Seurat object
}
\description{
Remove specific cell types based on marker gene expression
}
