% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieGEhistogram}
\alias{wookieGEhistogram}
\title{Compare Log Normalisation and SCT (Histogram)}
\usage{
wookieGEhistogram(seurat_obj, silentwookie = FALSE)
}
\arguments{
\item{seurat_obj}{Seurat Object with both RNA and SCT assays}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
Histogram comparing Log Normalisation and SCT
}
\description{
Compare Log Normalisation and SCT
}
