% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieGetPC}
\alias{wookieGetPC}
\title{Get Optimal Number of PCs}
\usage{
wookieGetPC(seurat_obj, reduction = "pca", silentwookie = FALSE)
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{reduction}{Type of reduction (e.g., 'pca')}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
Number of PCs to use
}
\description{
Get the optimal number of principal components to use
}
