% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieMatrixQc}
\alias{wookieMatrixQc}
\title{Plot QC metrics of a sparse matrix}
\usage{
wookieMatrixQc(
  count_matrix_sparse,
  fill_color = "#589FFF",
  title = "",
  silentwookie = FALSE
)
}
\arguments{
\item{count_matrix_sparse}{Sparse matrix}

\item{fill_color}{Plot color (default: #589FFF)}

\item{title}{Title for the plot}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
QC plot
}
\description{
Plot QC metrics of a sparse matrix
}
