% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookiePcePlot}
\alias{wookiePcePlot}
\title{Function to plot expression per cluster}
\usage{
wookiePcePlot(
  seurat,
  seurat_clusters = "seurat_clusters",
  saveplot = FALSE,
  dpi = 1080,
  height = 10,
  width = 10,
  units = "cm",
  silentwookie = FALSE
)
}
\arguments{
\item{seurat_clusters}{default is 'seurat_clusters'}

\item{saveplot}{save a jpeg image to working directory}

\item{dpi}{default is 1080}

\item{height}{default is 10}

\item{width}{default is 10}

\item{units}{deafults is 'cm'}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}

\item{seurat_obj}{Seurat object}
}
\value{
pce_plot
}
\description{
Function to plot Cluster relationship Tree
}
