% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieSilhouettePlot}
\alias{wookieSilhouettePlot}
\title{Function to plot silhouette scores for each cluster}
\usage{
wookieSilhouettePlot(
  seurat,
  cluster = "seurat_clusters",
  dims = 1:30,
  reduction = "pca",
  silentwookie = FALSE
)
}
\arguments{
\item{seurat}{Seurat object}

\item{cluster}{default is 'seurat_clusters'}

\item{dims}{dimension, default is 1:30}

\item{reduction}{default is 'pca'}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
silhouette plot
}
\description{
Function to plot silhouette scores for each cluster
}
