% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieUMAP3D}
\alias{wookieUMAP3D}
\title{Plot 3D Umap}
\usage{
wookieUMAP3D(
  seu,
  group.by = "seurat_clusters",
  palette = NULL,
  marker.size = 2,
  marker.width = 2,
  reduction = "umap",
  silentwookie = TRUE
)
}
\arguments{
\item{seu}{Seurat Object}

\item{group.by}{metadata column to group by}

\item{palette}{color palette to plot, Default is NULL}

\item{marker.size}{default 2}

\item{marker.width}{default 2}

\item{reduction}{default is umap}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}
}
\value{
plot
}
\description{
Plot 3D UMAP
}
