% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieUmapWizard}
\alias{wookieUmapWizard}
\title{Plot UMAPs to test features, min.dist values, and dimensions}
\usage{
wookieUmapWizard(
  object = seu_obj,
  features = NULL,
  min.dist = 0.3,
  max_features = 3000,
  ftype = "HVG",
  step = 500,
  out_name = "combined_umap",
  silentwookie = FALSE,
  mode = "features",
  reduction = "pca",
  max_dims = 30,
  dims_step = 5,
  nn_min = 10,
  nn_max = 50,
  nn_step = 10
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to use for UMAP (required for feature mode)}

\item{min.dist}{Minimum distance parameter for UMAP}

\item{max_features}{Maximum number of features to test (for feature mode)}

\item{ftype}{Feature type ('HVG' for highly variable genes or 'MAG' for most abundant genes)}

\item{step}{Step size for incrementing the number of features (for feature mode)}

\item{out_name}{Name to assign to the combined plot}

\item{silentwookie}{Stop Wookie from printing puns, default is FALSE}

\item{mode}{Mode of operation: 'features', 'min.dist', or 'dims'}

\item{reduction}{Reduction method for UMAP (for min.dist and dims modes)}

\item{max_dims}{Maximum number of dimensions to test (for dims mode)}

\item{dims_step}{Step size for incrementing the number of dimensions (for dims mode)}
}
\value{
Combined UMAP plot
}
\description{
Plot multiple UMAPs to test features, min.dist values, or dimensions
}
