% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{get_optimal_pcs}
\alias{get_optimal_pcs}
\title{get_optimal_pcs}
\usage{
get_optimal_pcs(seurat_object, reduction = "pca")
}
\arguments{
\item{seurat_obj}{Seurat Object}
}
\value{
number of PC's to use
}
\description{
Function to get the optimal number of principal components to use ...
}
