% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{multi_f_plots}
\alias{multi_f_plots}
\title{Plot multiple features}
\usage{
multi_f_plots(seuratObject, featureList, ncol = 3, pt.size = 0.8)
}
\arguments{
\item{featureList}{list of features to plot ...}
}
\value{
plot saved to global environment
}
\description{
plots n number features given in a list with custom colour scale ...
}
