% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_compare_normalisation}
\alias{plot_compare_normalisation}
\title{plot_compare_normalisation}
\usage{
plot_compare_normalisation(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat Object with both RNA and SCT assays}
}
\value{
plot
}
\description{
Function to compare Log Normalisation and SCT, Ensure Seurat Obj is SCTransformed and both scale.data contains all genes ...
}
