% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_compare_normalisation}
\alias{plot_compare_normalisation}
\title{Compare Normalization Methods}
\usage{
plot_compare_normalisation(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object with both RNA and SCT assays}
}
\value{
A plot comparing different normalization methods
}
\description{
Compare raw counts, normalized counts, SCTransform counts, and scaled data
}
