% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_gene_expression_histogram}
\alias{plot_gene_expression_histogram}
\title{plot_gene_expression_histogram}
\usage{
plot_gene_expression_histogram(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat Object with both RNA and SCT assays}
}
\value{
histogram
}
\description{
Function to compare Log Normalisation and SCT, Ensure Seurat Obj is SCTransformed and both scale.data contains all genes ...
}
