% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{plot_multi_feat_umap}
\alias{plot_multi_feat_umap}
\title{Plot UMAPs to test features}
\usage{
plot_multi_feat_umap(
  object = seu_obj,
  features = features,
  min.dist = 0.1,
  max_features = 3000,
  ftype = "HVG",
  step = 500,
  out_name = "combined_umap"
)
}
\value{
plot saved to global environment
}
\description{
plot multiple UMAP's for different numbers of a feature i.e Highly variable genes or Most Abundant genes ...
}
