% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_dotplot}
\alias{wookie_dotplot}
\title{Plot dotplots with color map}
\usage{
wookie_dotplot(
  seurat_obj,
  feature_list,
  assay = "RNA",
  scale.by = "size",
  tag = "DotPlot"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{feature_list}{List of markers}

\item{assay}{default RNA}

\item{scale.by}{scale.by}

\item{tag}{plot title}
}
\value{
Pdotplot
}
\description{
dotplot
}
