% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookie_matrix_qc_plot}
\alias{wookie_matrix_qc_plot}
\title{wookie_matrix_qc_plot}
\usage{
wookie_matrix_qc_plot(count_matrix_sparse, fill_color = "#589FFF")
}
\arguments{
\item{count_matrix_sparse}{sparce matrix}

\item{fill_color}{plot color, default is #589FFF ...}
}
\value{
qc plot
}
\description{
Function to to plot qc metrics of a sparce matrix ...
}
