% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieqc}
\alias{wookieqc}
\alias{wookieQc}
\title{Seurat Object Quality Control function}
\usage{
wookieQc(
  seurat_obj,
  nf_min = 0,
  nf_max = 20000,
  nc_max = 2e+05,
  nc_min = 0,
  pmt = 20,
  ptr_max = NULL,
  ptr_min = NULL,
  species = "Mouse",
  pt.size = NULL,
  legend = TRUE,
  calculateMetrics = TRUE,
  silentwookie = FALSE
)
}
\arguments{
\item{nf_min}{Minimum number of features ...}

\item{nf_max}{Maximum number of features ...}

\item{pmt}{Percentage of mitochondrial genes ...}

\item{ptr_max}{maximum percentage of ribosomal genes ...}

\item{ptr_min}{minimum percentage of ribosomal genes ...}

\item{species}{species in dataset Mouse or Human only ...}

\item{pt.size}{data points in violin plot}

\item{silentwookie}{stop wookie from printing puns, default is FALSE}

\item{matrix}{Seurat object ...}

\item{nc}{Maximum number of counts ...}

\item{group}{Grouping variable ...}

\item{colors}{Colors for facetting ...}
}
\value{
Seurat object after quality control
}
\description{
Run Iteratively for the necessary QC...
}
