% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WookieTools.R
\name{wookieqc}
\alias{wookieqc}
\alias{wookie_qc}
\title{Seurat Object Quality Control function}
\usage{
wookie_qc(
  matrix,
  nf_min = 0,
  nf_max = 20000,
  nc = 2e+05,
  pmt = 20,
  ptr = 100,
  group = "orig.ident",
  colors = NULL
)
}
\arguments{
\item{matrix}{Seurat object ...}

\item{nf_min}{Minimum number of features ...}

\item{nf_max}{Maximum number of features ...}

\item{nc}{Maximum number of counts ...}

\item{pmt}{Percentage of mitochondrial genes ...}

\item{ptr}{Percentage of ribosomal genes ...}

\item{group}{Grouping variable ...}

\item{colors}{Colors for facetting ...}
}
\value{
Seurat object after quality control
}
\description{
Run Iteratively for the necessary QC...
}
